Crawing Functionality for specific Item added and tested
