+++ src/main/java/IWebCrawler.java
@@ -1,4 +1,5 @@
 import Entities.Item;
+import org.jsoup.nodes.Document;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -8,7 +9,7 @@ public interface IWebCrawler {
     String baseURL = null;
 
 
-    public ArrayList<Item> craw(String baseURL) throws IOException;
+    public ArrayList<Document> craw(String baseURL) throws IOException;
 
 
 }
